Move active_model around.
