+++ config/initializers/register_adapters.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-require 'valkyrie/active_model'
+require 'valkyrie'
 Rails.application.config.to_prepare do
   Valkyrie::Adapter.register(
     Valkyrie::Persistence::Postgres,
